+++ stories/addon/cc-header-addon.stories.js
@@ -48,6 +48,9 @@ export default {
 const conf = {
   component: 'cc-header-addon',
   css: `
+    :host {
+      max-width: 82em !important;
+    }
     cc-header-addon {
       margin-bottom: 1rem;
     }